Fixed tile visual glitch by truncating tile size decims

+++ Pathfinding/build.cpp
@@ -9,7 +9,7 @@ constexpr int MAP_WIDTH = 40;				// Map stuff
 constexpr int MAP_HEIGHT = 40;
 constexpr float SIZE_MULTIPLIER = 1.f;		// Tile
 constexpr int DECREASE_BOX_SIZE = 5;
-constexpr float DEFAULT_TILE_SIZE = static_cast<float>(WINDOW_WIDTH) / (MAP_WIDTH + 2 + DECREASE_BOX_SIZE) * SIZE_MULTIPLIER;
+constexpr float DEFAULT_TILE_SIZE = static_cast<float>(static_cast<int>(WINDOW_WIDTH)/(MAP_WIDTH+2+DECREASE_BOX_SIZE) * SIZE_MULTIPLIER);
 constexpr int SEARCHED_DISPLAY_RATE = 1;	// How fast nodes are displayed ( Higher value == slower )
 constexpr int PATH_CONSTRUCTION_RATE = 15;
